Refactor colormodes logic to its own module
