+++ docs/api/internal/colormodes.rst
@@ -1,0 +1,7 @@
+ridgeplot._colormodes
+=================
+
+Colormode utilities.
+
+.. automodule:: ridgeplot._colormodes
+   :private-members:
