+++ src/ridgeplot/_figure_factory.py
@@ -1,11 +1,18 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Literal, Protocol, cast
+from typing import TYPE_CHECKING, cast
 
 from plotly import graph_objects as go
 
-from ridgeplot._colors import apply_alpha, get_color, get_colorscale, validate_colorscale
+from ridgeplot._colormodes import (
+    COLORMODE_MAPS,
+    Colormode,
+    ColorsArray,
+    MidpointsContext,
+    pre_compute_colors,
+)
+from ridgeplot._colors import get_colorscale, validate_colorscale
 from ridgeplot._types import (
     CollectionL1,
     CollectionL2,
@@ -15,7 +22,7 @@ from ridgeplot._types import (
 )
 from ridgeplot._utils import (
     get_collection_array_shape,
-    normalise_min_max,
+    get_xy_extrema,
     normalise_row_attrs,
     ordered_dedup,
 )
@@ -49,18 +56,6 @@ Example
 >>> labels_array: ShallowLabelsArray = ["trace 1", "trace 2", "trace 3"]
 """
 
-ColorsArray = CollectionL2[str]
-"""A :data:`ColorsArray` represents the colors of traces in a ridgeplot.
-
-Example
--------
-
->>> colors_array: ColorsArray = [
-...     ["red", "blue", "green"],
-...     ["orange", "purple"],
-... ]
-"""
-
 ShallowColorsArray = CollectionL1[str]
 """Shallow type for :data:`ColorsArray`.
 
@@ -70,22 +65,6 @@ Example
 >>> colors_array: ShallowColorsArray = ["red", "blue", "green"]
 """
 
-MidpointsArray = CollectionL2[float]
-"""A :data:`MidpointsArray` represents the midpoints of colorscales in a
-ridgeplot.
-
-Example
--------
-
->>> midpoints_array: MidpointsArray = [
-...     [0.2, 0.5, 1],
-...     [0.3, 0.7],
-... ]
-"""
-
-Colormode = Literal["row-index", "trace-index", "trace-index-row-wise", "mean-minmax", "mean-means"]
-"""The :paramref:`ridgeplot.ridgeplot.colormode` argument in
-:func:`ridgeplot.ridgeplot()`."""
 
 _D3HF = ".7"
 """Default (d3-format) format for floats in hover labels.
@@ -107,56 +86,6 @@ See :func:`draw_density_trace`.
 """
 
 
-def get_xy_extrema(densities: Densities) -> tuple[Numeric, Numeric, Numeric, Numeric]:
-    r"""Get the global x-y extrema (x_min, x_max, y_min, y_max) over all
-    :data:`~ridgeplot._types.DensityTrace`\s in the
-    :data:`~ridgeplot._types.Densities` array.
-
-    Parameters
-    ----------
-    densities
-        A :data:`~ridgeplot._types.Densities` array.
-
-
-    Returns
-    -------
-    Tuple[Numeric, Numeric, Numeric, Numeric]
-        A tuple of the form (x_min, x_max, y_min, y_max).
-
-    Examples
-    --------
-    >>> get_xy_extrema(
-    ...     [
-    ...         [
-    ...             [(0, 0), (1, 1), (2, 2), (3, 3)],
-    ...             [(0, 0), (1, 1), (2, 2)],
-    ...             [(0, 0), (1, 1), (2, 2), (3, 3), (4, 4)],
-    ...         ],
-    ...         [
-    ...             [(-2, 2), (-1, 1), (0, 1)],
-    ...             [(2, 2), (3, 1), (4, 1)],
-    ...         ],
-    ...     ]
-    ... )
-    (-2, 4, 0, 4)
-    """
-    if len(densities) == 0:
-        raise ValueError("The densities array should not be empty.")
-    x_flat: list[Numeric] = []
-    y_flat: list[Numeric] = []
-    for row in densities:
-        for trace in row:
-            for x, y in trace:
-                x_flat.append(x)
-                y_flat.append(y)
-    return min(x_flat), max(x_flat), min(y_flat), max(y_flat)
-
-
-def _mul(a: tuple[Numeric, ...], b: tuple[Numeric, ...]) -> tuple[Numeric, ...]:
-    """Multiply two tuples element-wise."""
-    return tuple(a_i * b_i for a_i, b_i in zip(a, b))
-
-
 @dataclass
 class RidgeplotTrace:
     trace: DensityTrace
@@ -168,99 +97,6 @@ class RidgeplotTrace:
 class RidgeplotRow:
     traces: list[RidgeplotTrace]
     y_shifted: float
-
-
-@dataclass
-class MidpointsContext:
-    densities: Densities
-    n_rows: int
-    n_traces: int
-    x_min: Numeric
-    x_max: Numeric
-
-
-class MidpointsFunc(Protocol):
-    def __call__(self, ctx: MidpointsContext) -> MidpointsArray: ...
-
-
-def _compute_midpoints_row_index(ctx: MidpointsContext) -> MidpointsArray:
-    return [
-        [((ctx.n_rows - 1) - ith_row) / (ctx.n_rows - 1)] * len(row)
-        for ith_row, row in enumerate(ctx.densities)
-    ]
-
-
-def _compute_midpoints_trace_index(ctx: MidpointsContext) -> MidpointsArray:
-    midpoints = []
-    ith_trace = 0
-    for row in ctx.densities:
-        midpoints_row = []
-        for _ in row:
-            midpoints_row.append(((ctx.n_traces - 1) - ith_trace) / (ctx.n_traces - 1))
-            ith_trace += 1
-        midpoints.append(midpoints_row)
-    return midpoints
-
-
-def _compute_midpoints_trace_index_row_wise(ctx: MidpointsContext) -> MidpointsArray:
-    return [
-        [((len(row) - 1) - ith_row_trace) / (len(row) - 1) for ith_row_trace in range(len(row))]
-        for row in ctx.densities
-    ]
-
-
-def _compute_midpoints_mean_minmax(ctx: MidpointsContext) -> MidpointsArray:
-    midpoints = []
-    for row in ctx.densities:
-        midpoints_row = []
-        for trace in row:
-            x, y = zip(*trace)
-            midpoints_row.append(
-                normalise_min_max(sum(_mul(x, y)) / sum(y), min_=ctx.x_min, max_=ctx.x_max)
-            )
-        midpoints.append(midpoints_row)
-    return midpoints
-
-
-def _compute_midpoints_mean_means(ctx: MidpointsContext) -> MidpointsArray:
-    means = []
-    for row in ctx.densities:
-        means_row = []
-        for trace in row:
-            x, y = zip(*trace)
-            means_row.append(sum(_mul(x, y)) / sum(y))
-        means.append(means_row)
-    min_mean = min([min(row) for row in means])
-    max_mean = max([max(row) for row in means])
-    return [
-        [normalise_min_max(mean, min_=min_mean, max_=max_mean) for mean in row] for row in means
-    ]
-
-
-def pre_compute_colors(
-    colorscale: ColorScale,
-    colormode: Colormode,
-    coloralpha: float | None,
-    midpoints_context: MidpointsContext,
-) -> ColorsArray:
-    def _get_color(mp: float) -> str:
-        color = get_color(colorscale, midpoint=mp)
-        if coloralpha is not None:
-            color = apply_alpha(color, alpha=coloralpha)
-        return color
-
-    midpoints_func = COLORMODE_MAPS[colormode]
-    midpoints = midpoints_func(ctx=midpoints_context)
-    return [[_get_color(midpoint) for midpoint in row] for row in midpoints]
-
-
-COLORMODE_MAPS: dict[Colormode, MidpointsFunc] = {
-    "row-index": _compute_midpoints_row_index,
-    "trace-index": _compute_midpoints_trace_index,
-    "trace-index-row-wise": _compute_midpoints_trace_index_row_wise,
-    "mean-minmax": _compute_midpoints_mean_minmax,
-    "mean-means": _compute_midpoints_mean_means,
-}
 
 
 def draw_base(
