+++ tox.ini
@@ -50,7 +50,7 @@ commands =
     coverage-combine: mv .coverage {env:COVERAGE_FILE}
     coverage-combine: coverage xml  --data-file={env:COVERAGE_FILE} --omit="cicd_utils/*,tests/*" -o {env:COVERAGE_XML_FILE}
     coverage-combine: coverage html --data-file={env:COVERAGE_FILE} --omit="cicd_utils/*,tests/*"
-    coverage-combine: coverage report --data-file={env:COVERAGE_FILE} --omit="cicd_utils/*,tests/*" --fail-under=95
+    coverage-combine: coverage report --data-file={env:COVERAGE_FILE} --omit="cicd_utils/*,tests/*" --fail-under=98
     coverage-combine: diff-cover {env:_DIFFCOVER_DFLT_ARGS}
 [testenv:pre-commit]
 description = run code style and linting checks with pre-commit
